+++ .vscode/advent_of_code.code-workspace
@@ -1,0 +1,16 @@
+{
+	"folders": [
+		{
+			"path": ".."
+		}
+	],
+	"settings": {
+		"jest.disabledWorkspaceFolders": [
+			"day_1",
+			"day_2",
+			"day_5",
+			"day_4",
+			"day_3"
+		]
+	}
+}
