+++ .vscode/launch.json
@@ -13,6 +13,25 @@
       ],
       "program": "${file}",
       "cwd": "${fileDirname}"
+    },
+    {
+      "type": "node",
+      "name": "vscode-jest-tests.v2.advent_of_code",
+      "request": "launch",
+      "args": [
+        "jest",
+        "--runInBand",
+        "--watchAll=false",
+        "--testNamePattern",
+        "${jest.testNamePattern}",
+        "--runTestsByPath",
+        "${jest.testFile}"
+      ],
+      "cwd": "${workspaceFolder}",
+      "console": "integratedTerminal",
+      "internalConsoleOptions": "neverOpen",
+      "disableOptimisticBPs": true,
+      "program": "${workspaceFolder}/npx"
     }
   ]
 }
