+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "jest.jestCommandLine": "npx jest"
+}
